refactor(desktop): extract reminder text formatting
