+++ apps/desktop/src/services/AgentService/task/center.ts
@@ -1,13 +1,15 @@
 // Copyright (c) 2026. Qian Cheng. Licensed under GPL v3
 
-import { getMarkdown } from 'markstream-vue';
-
-import { getLocale, tt } from '@/i18n';
+import { tt } from '@/i18n';
 import { eventService } from '@/services/EventService';
 import { AppEvent, type SessionStatusReminderPayload } from '@/services/EventService/types';
-import type { PendingToolApproval, SessionMessage } from '@/types/session';
+import {
+    buildWaitingApprovalBody,
+    buildWaitingUserQuestionBody,
+    summarizeLatestAssistantResponse,
+    summarizeNotificationText,
+} from '@/utils/reminderText';
 import { getSessionStatusReminderContent } from '@/utils/session';
-import { collapseWhitespace } from '@/utils/text';
 
 import { AiError, AiErrorCode } from '../contracts/errors';
 import type { ConversationRuntimeEnvironment, TurnEvent } from '../execution';
@@ -41,99 +43,6 @@ interface MutableSessionTask {
 }
 
 const TERMINAL_TASK_RETENTION_MS = 5 * 60 * 1000;
-const STATUS_REMINDER_MAX_BODY_CHARS = 220;
-const STATUS_REMINDER_MAX_COMMAND_CHARS = 160;
-const STATUS_REMINDER_MAX_SUMMARY_LINES = 4;
-const REMINDER_MARKDOWN_ESCAPE_PATTERN = /\\([\\`*_{}[\]()#+.!>-])/g;
-const REMINDER_PATH_LIKE_TOKEN_PATTERN = /\S*[\\/]\S*/g;
-const REMINDER_PATH_WRAPPER_TRIM_PATTERN = /^[("'[{]+|[)"'\],.;:!?}]+$/g;
-const REMINDER_PATH_SEGMENT_PATTERN = '(?:[A-Za-z0-9._@-]+|\\*{1,2})';
-const REMINDER_POSIX_RELATIVE_PATH_PATTERN = new RegExp(
-    `^(?:${REMINDER_PATH_SEGMENT_PATTERN}/)+${REMINDER_PATH_SEGMENT_PATTERN}$`
-);
-const REMINDER_WINDOWS_RELATIVE_PATH_PATTERN = new RegExp(
-    `^(?:${REMINDER_PATH_SEGMENT_PATTERN}\\\\)+${REMINDER_PATH_SEGMENT_PATTERN}$`
-);
-const REMINDER_POSIX_ABSOLUTE_PATH_PATTERN = new RegExp(
-    `^(?:/|\\.{1,2}/|~/)(?:${REMINDER_PATH_SEGMENT_PATTERN}/)*${REMINDER_PATH_SEGMENT_PATTERN}$`
-);
-const REMINDER_WINDOWS_ABSOLUTE_PATH_PATTERN = new RegExp(
-    `^(?:[A-Za-z]:\\\\|\\.{1,2}\\\\|~\\\\)(?:${REMINDER_PATH_SEGMENT_PATTERN}\\\\)*${REMINDER_PATH_SEGMENT_PATTERN}$`
-);
-const REMINDER_WINDOWS_UNC_PATH_PATTERN = new RegExp(
-    `^\\\\\\\\${REMINDER_PATH_SEGMENT_PATTERN}(?:\\\\${REMINDER_PATH_SEGMENT_PATTERN})+$`
-);
-
-type ReminderTextMode = 'natural' | 'command' | 'summary';
-// markstream emits standard markdown-it tokens and may also surface a custom
-// inline `link` token with pre-flattened label text.
-type ReminderMarkdownBaseToken = {
-    type: string;
-    tag?: string;
-    content?: string;
-    markup?: string;
-    children?: ReminderMarkdownToken[] | null;
-};
-type ReminderMarkdownLinkToken = ReminderMarkdownBaseToken & {
-    type: 'link';
-    text?: string;
-};
-type ReminderMarkdownToken = ReminderMarkdownBaseToken | ReminderMarkdownLinkToken;
-type ReminderInlineHtmlTag = {
-    hasAttributes: boolean;
-    isClosing: boolean;
-    isSelfClosing: boolean;
-    tagName: string;
-};
-
-const REMINDER_INLINE_LINE_BREAK_HTML_TAGS = new Set(['br', 'hr']);
-const REMINDER_STRIPPABLE_INLINE_HTML_TAGS = new Set([
-    'a',
-    'abbr',
-    'b',
-    'bdi',
-    'bdo',
-    'cite',
-    'code',
-    'data',
-    'del',
-    'dfn',
-    'div',
-    'em',
-    'i',
-    'ins',
-    'kbd',
-    'li',
-    'mark',
-    'p',
-    'pre',
-    'q',
-    'rp',
-    'rt',
-    'ruby',
-    's',
-    'samp',
-    'small',
-    'span',
-    'strong',
-    'sub',
-    'sup',
-    'time',
-    'u',
-    'ul',
-    'ol',
-    'var',
-    'wbr',
-]);
-
-const reminderMarkdownParser = getMarkdown('touchai-reminder-markdown', {
-    enableContainers: false,
-    markdownItOptions: {
-        breaks: true,
-    },
-});
-
-/** 深拷贝任务快照，确保外部订阅者无法直接修改内部状态。 */
 function cloneTaskSnapshot(snapshot: SessionTaskSnapshot): SessionTaskSnapshot {
     return cloneTaskValue(snapshot);
 }
@@ -160,569 +69,6 @@ function isTerminalStatus(status: SessionTaskSnapshot['status']): boolean {
     return status === 'completed' || status === 'failed' || status === 'cancelled';
 }
 
-function truncateNotificationText(value: string, maxChars: number): string {
-    if (value.length <= maxChars) {
-        return value;
-    }
-
-    return `${value.slice(0, maxChars - 3).trimEnd()}...`;
-}
-
-function isEnglishReminderLocale(): boolean {
-    return getLocale() === 'en-US';
-}
-
-function getReminderListSeparator(): string {
-    return isEnglishReminderLocale() ? ', ' : '、';
-}
-
-function getReminderClauseSeparator(): string {
-    return isEnglishReminderLocale() ? '; ' : '；';
-}
-
-function getReminderSentenceSeparator(): string {
-    return isEnglishReminderLocale() ? '. ' : '。';
-}
-
-function getReminderColonSeparator(): string {
-    return isEnglishReminderLocale() ? ': ' : '：';
-}
-
-function hasTerminalPunctuation(value: string): boolean {
-    return /[.!?。！？；;:：]$/.test(value.trim());
-}
-
-/** Rehydrate backslash-escaped markdown so token parsing sees the intended text. */
-function unescapeReminderMarkdown(value: string): string {
-    return value.replace(REMINDER_MARKDOWN_ESCAPE_PATTERN, '$1');
-}
-
-/** Identify tokens that look like filesystem paths or globs rather than markdown escapes. */
-function isReminderPathLikeToken(token: string): boolean {
-    const core = token.replace(REMINDER_PATH_WRAPPER_TRIM_PATTERN, '');
-    return (
-        REMINDER_POSIX_RELATIVE_PATH_PATTERN.test(core) ||
-        REMINDER_WINDOWS_RELATIVE_PATH_PATTERN.test(core) ||
-        REMINDER_POSIX_ABSOLUTE_PATH_PATTERN.test(core) ||
-        REMINDER_WINDOWS_ABSOLUTE_PATH_PATTERN.test(core) ||
-        REMINDER_WINDOWS_UNC_PATH_PATTERN.test(core)
-    );
-}
-
-/** Escape path-like fragments so markdown emphasis markers inside file paths are preserved. */
-function protectPathLikeMarkdownToken(token: string): string {
-    if (!isReminderPathLikeToken(token)) {
-        return token;
-    }
-
-    if (!/[`*_]/.test(token)) {
-        return token;
-    }
-
-    if (/^!?\[[^\]]*]\([^)]+\)$/.test(token)) {
-        return token;
-    }
-
-    return token.replace(/([\\`*_])/g, '\\$1');
-}
-
-/** Normalize reminder input before markdown parsing and protect path and glob syntax. */
-function prepareReminderMarkdownSource(value: string): string {
-    const normalized = value.replace(/\r\n?/g, '\n');
-    const protectedPaths: string[] = [];
-    const withPlaceholders = normalized.replace(REMINDER_PATH_LIKE_TOKEN_PATTERN, (token) => {
-        if (!isReminderPathLikeToken(token)) {
-            return token;
-        }
-
-        const placeholder = `%%TOUCHAI_REMINDER_PATH_${protectedPaths.length}%%`;
-        protectedPaths.push(protectPathLikeMarkdownToken(token));
-        return placeholder;
-    });
-    const unescaped = unescapeReminderMarkdown(withPlaceholders);
-    return protectedPaths.reduce(
-        (text, token, index) => text.replace(`%%TOUCHAI_REMINDER_PATH_${index}%%`, token),
-        unescaped
-    );
-}
-
-/** Reduce block HTML to plain text for notification-safe summaries. */
-function stripHtmlToText(value: string): string {
-    if (!value) {
-        return '';
-    }
-
-    if (typeof DOMParser !== 'undefined') {
-        try {
-            return new DOMParser().parseFromString(value, 'text/html').body.textContent ?? '';
-        } catch {
-            // Fall back to a conservative tag strip when DOM parsing is unavailable.
-        }
-    }
-
-    return value.replace(/<[^>]+>/g, ' ');
-}
-
-function parseReminderInlineHtmlTag(value: string): ReminderInlineHtmlTag | null {
-    const match = value.match(/^<\s*(\/)?\s*([A-Za-z][A-Za-z0-9:-]*)([\s\S]*?)>$/);
-    if (!match) {
-        return null;
-    }
-
-    const suffix = match[3] ?? '';
-    const normalizedSuffix = suffix.replace(/\/\s*$/, '');
-    return {
-        hasAttributes: /\S/.test(normalizedSuffix),
-        isClosing: Boolean(match[1]),
-        isSelfClosing: /\/\s*$/.test(suffix),
-        tagName: match[2]?.toLowerCase() ?? '',
-    };
-}
-
-function hasMatchingReminderInlineClosingTag(
-    tokens: ReminderMarkdownToken[],
-    startIndex: number,
-    tagName: string
-): boolean {
-    let depth = 0;
-    for (let index = startIndex + 1; index < tokens.length; index += 1) {
-        const token = tokens[index];
-        if (token?.type !== 'html_inline') {
-            continue;
-        }
-
-        const parsedTag = parseReminderInlineHtmlTag(token.content ?? '');
-        if (!parsedTag || parsedTag.tagName !== tagName) {
-            continue;
-        }
-
-        if (!parsedTag.isClosing) {
-            depth += 1;
-            continue;
-        }
-
-        if (depth === 0) {
-            return true;
-        }
-
-        depth -= 1;
-    }
-
-    return false;
-}
-
-function normalizeReminderFallbackHtml(value: string): string {
-    let normalized = value
-        .replace(/<br\s*\/?>/gi, '\n')
-        .replace(/<hr\s*\/?>/gi, '\n')
-        .replace(/<wbr\s*\/?>/gi, '');
-
-    for (const tagName of REMINDER_STRIPPABLE_INLINE_HTML_TAGS) {
-        if (REMINDER_INLINE_LINE_BREAK_HTML_TAGS.has(tagName)) {
-            continue;
-        }
-
-        normalized = normalized.replace(new RegExp(`</?${tagName}\\b[^>]*>`, 'gi'), '');
-    }
-
-    return normalized;
-}
-
-function normalizeReminderInlineHtmlToken(
-    tokenContent: string,
-    tokens: ReminderMarkdownToken[],
-    index: number,
-    openTags: Map<string, number>
-): string {
-    const parsedTag = parseReminderInlineHtmlTag(tokenContent);
-    if (!parsedTag) {
-        return tokenContent;
-    }
-
-    if (REMINDER_INLINE_LINE_BREAK_HTML_TAGS.has(parsedTag.tagName)) {
-        return '\n';
-    }
-
-    const shouldStripTag =
-        REMINDER_STRIPPABLE_INLINE_HTML_TAGS.has(parsedTag.tagName) || parsedTag.hasAttributes;
-    if (!shouldStripTag) {
-        return tokenContent;
-    }
-
-    if (parsedTag.isClosing) {
-        const openCount = openTags.get(parsedTag.tagName) ?? 0;
-        if (openCount <= 0) {
-            return tokenContent;
-        }
-
-        if (openCount === 1) {
-            openTags.delete(parsedTag.tagName);
-        } else {
-            openTags.set(parsedTag.tagName, openCount - 1);
-        }
-        return '';
-    }
-
-    if (parsedTag.isSelfClosing) {
-        return '';
-    }
-
-    if (!hasMatchingReminderInlineClosingTag(tokens, index, parsedTag.tagName)) {
-        return tokenContent;
-    }
-
-    openTags.set(parsedTag.tagName, (openTags.get(parsedTag.tagName) ?? 0) + 1);
-    return '';
-}
-
-/** Keep command previews copyable by converting typographic quotes back to ASCII. */
-function normalizeCommandTypography(value: string): string {
-    return value.replace(/[“”]/g, '"').replace(/[‘’]/g, "'");
-}
-
-/** Keep approval command previews recognizable instead of markdown-cleaned. */
-function summarizeCommandPreview(
-    value: string | null | undefined,
-    maxChars: number
-): string | null {
-    const normalized = collapseWhitespace(
-        normalizeCommandTypography((value ?? '').replace(/\r\n?/g, '\n'))
-    );
-    if (!normalized) {
-        return null;
-    }
-
-    return truncateNotificationText(normalized, maxChars);
-}
-
-/** Flatten inline markdown tokens into readable plain text for reminder clauses. */
-function extractReminderInlineText(
-    tokens: ReminderMarkdownToken[] | null | undefined,
-    fallback: string
-): string {
-    if (!tokens?.length) {
-        return fallback;
-    }
-
-    let text = '';
-    const openInlineHtmlTags = new Map<string, number>();
-    for (let index = 0; index < tokens.length; index += 1) {
-        const token = tokens[index];
-        if (!token) {
-            continue;
-        }
-
-        switch (token.type) {
-            case 'text':
-            case 'code_inline':
-                text += token.content ?? '';
-                break;
-            case 'softbreak':
-            case 'hardbreak':
-                text += '\n';
-                break;
-            case 'html_inline':
-                text += normalizeReminderInlineHtmlToken(
-                    token.content ?? '',
-                    tokens,
-                    index,
-                    openInlineHtmlTags
-                );
-                break;
-            case 'link':
-                text += extractReminderInlineText(
-                    token.children,
-                    (token as ReminderMarkdownLinkToken).text ?? token.content ?? ''
-                );
-                break;
-            case 'link_open':
-            case 'link_close':
-                break;
-            default:
-                if (token.children?.length) {
-                    text += extractReminderInlineText(token.children, '');
-                    break;
-                }
-
-                if (token.type.endsWith('_open') || token.type.endsWith('_close')) {
-                    break;
-                }
-
-                text += token.content ?? '';
-                break;
-        }
-    }
-
-    return text || fallback;
-}
-
-/** Split normalized text into non-empty clauses for later summary joining. */
-function pushReminderClauses(target: string[], value: string): void {
-    for (const line of value.split('\n')) {
-        const clause = collapseWhitespace(line);
-        if (clause) {
-            target.push(clause);
-        }
-    }
-}
-
-/** Provide a conservative fallback when markdown tokenization fails. */
-function fallbackReminderClauses(value: string): string[] {
-    const clauses: string[] = [];
-    const fallbackText = normalizeReminderFallbackHtml(value);
-    pushReminderClauses(clauses, fallbackText);
-    return clauses;
-}
-
-/** Parse reminder markdown into plain-text clauses, including tables and code blocks. */
-function collectReminderClauses(
-    value: string,
-    mode: Extract<ReminderTextMode, 'natural' | 'summary'>
-): string[] {
-    const source = prepareReminderMarkdownSource(value);
-    if (!source.trim()) {
-        return [];
-    }
-
-    try {
-        const tokens = reminderMarkdownParser.parse(source, {}) as ReminderMarkdownToken[];
-        const clauses: string[] = [];
-        let currentTableRow: string[] | null = null;
-        let insideTableCell = false;
-
-        for (const token of tokens) {
-            switch (token.type) {
-                case 'tr_open':
-                    currentTableRow = [];
-                    break;
-                case 'tr_close': {
-                    const row = currentTableRow?.filter(Boolean).join(getReminderListSeparator());
-                    if (row) {
-                        clauses.push(row);
-                    }
-                    currentTableRow = null;
-                    insideTableCell = false;
-                    break;
-                }
-                case 'th_open':
-                case 'td_open':
-                    insideTableCell = true;
-                    break;
-                case 'th_close':
-                case 'td_close':
-                    insideTableCell = false;
-                    break;
-                case 'inline': {
-                    const text = extractReminderInlineText(token.children, token.content ?? '');
-                    if (!text) {
-                        break;
-                    }
-
-                    if (insideTableCell && currentTableRow) {
-                        const cell = collapseWhitespace(text);
-                        if (cell) {
-                            currentTableRow.push(cell);
-                        }
-                        break;
-                    }
-
-                    pushReminderClauses(clauses, text);
-                    break;
-                }
-                case 'fence':
-                case 'code_block':
-                    pushReminderClauses(clauses, token.content ?? '');
-                    break;
-                case 'html_block':
-                    pushReminderClauses(clauses, stripHtmlToText(token.content ?? ''));
-                    break;
-                default:
-                    break;
-            }
-        }
-
-        if (mode === 'summary') {
-            return clauses.slice(0, STATUS_REMINDER_MAX_SUMMARY_LINES);
-        }
-
-        return clauses;
-    } catch {
-        const clauses = fallbackReminderClauses(source);
-        return mode === 'summary' ? clauses.slice(0, STATUS_REMINDER_MAX_SUMMARY_LINES) : clauses;
-    }
-}
-
-/** Identify short fragments that can be merged into a compact summary line. */
-function isShortReminderClause(value: string): boolean {
-    const trimmed = value.trim();
-    if (!trimmed || hasTerminalPunctuation(trimmed)) {
-        return false;
-    }
-
-    if (trimmed.includes(getReminderListSeparator().trim())) {
-        return false;
-    }
-
-    return trimmed.length <= (isEnglishReminderLocale() ? 32 : 24);
-}
-
-/** Join clauses without doubling separators after terminal punctuation. */
-function joinReminderSequence(clauses: string[], separator: string): string {
-    const [firstClause, ...restClauses] = clauses;
-    if (!firstClause) {
-        return '';
-    }
-
-    let result = firstClause;
-    for (const clause of restClauses) {
-        const joiner = hasTerminalPunctuation(result) ? ' ' : separator;
-        result = `${result}${joiner}${clause}`;
-    }
-
-    return result;
-}
-
-/** Build the final reminder sentence with locale-aware separators and summary shaping. */
-function joinReminderClauses(clauses: string[], mode: ReminderTextMode): string {
-    const uniqueClauses: string[] = [];
-    for (const clause of clauses) {
-        if (uniqueClauses[uniqueClauses.length - 1] === clause) {
-            continue;
-        }
-        uniqueClauses.push(clause);
-    }
-
-    if (uniqueClauses.length === 0) {
-        return '';
-    }
-
-    const [firstClause, ...restClauses] = uniqueClauses;
-    if (!firstClause) {
-        return '';
-    }
-
-    if (restClauses.length === 0) {
-        return firstClause;
-    }
-
-    const useTitledSummary =
-        mode === 'summary' &&
-        !hasTerminalPunctuation(firstClause) &&
-        firstClause.length <= (isEnglishReminderLocale() ? 60 : 40);
-    const separator =
-        mode === 'summary' && restClauses.every((clause) => isShortReminderClause(clause))
-            ? getReminderListSeparator()
-            : getReminderClauseSeparator();
-    const restText = joinReminderSequence(restClauses, separator);
-
-    if (!useTitledSummary) {
-        return joinReminderSequence(uniqueClauses, separator);
-    }
-
-    return `${firstClause}${getReminderColonSeparator()}${restText}`;
-}
-
-/** Append an extra reminder fragment while keeping the sentence readable. */
-function appendReminderClause(base: string, clause: string | null): string {
-    if (!clause) {
-        return base;
-    }
-
-    if (!base) {
-        return clause;
-    }
-
-    const separator = hasTerminalPunctuation(base) ? ' ' : getReminderSentenceSeparator();
-    return `${base}${separator}${clause}`;
-}
-
-/** Format a labeled reminder fragment such as a command preview. */
-function formatReminderLabelValue(label: string, value: string): string {
-    return `${label}${getReminderColonSeparator()}${value}`;
-}
-
-/** Convert markdown-rich content into a notification-ready plain-text summary. */
-function summarizeNotificationText(
-    value: string | null | undefined,
-    maxChars = STATUS_REMINDER_MAX_BODY_CHARS,
-    mode: ReminderTextMode = 'natural'
-) {
-    if (mode === 'command') {
-        return summarizeCommandPreview(value, maxChars);
-    }
-
-    const normalized = joinReminderClauses(collectReminderClauses(value ?? '', mode), mode);
-    if (!normalized) {
-        return null;
-    }
-
-    return truncateNotificationText(normalized, maxChars);
-}
-
-/** 从会话历史中提取最后一条 assistant 消息的摘要。 */
-function summarizeLatestAssistantResponse(history: SessionMessage[]): string | null {
-    for (let index = history.length - 1; index >= 0; index -= 1) {
-        const message = history[index];
-        if (message?.role !== 'assistant') {
-            continue;
-        }
-
-        const summary = summarizeNotificationText(
-            message.content,
-            STATUS_REMINDER_MAX_BODY_CHARS,
-            'summary'
-        );
-        if (summary) {
-            return summary;
-        }
-    }
-
-    return null;
-}
-
-/** 为等待审批状态构建通知正文，包含摘要和命令预览。 */
-function buildWaitingApprovalBody(approval: PendingToolApproval): string {
-    const summary =
-        summarizeNotificationText(approval.reason) ??
-        summarizeNotificationText(approval.description) ??
-        summarizeNotificationText(approval.title) ??
-        getSessionStatusReminderContent('waiting_approval');
-    const commandPreview = summarizeNotificationText(
-        approval.command,
-        STATUS_REMINDER_MAX_COMMAND_CHARS,
-        'command'
-    );
-
-    if (!commandPreview || commandPreview === summary) {
-        return summary;
-    }
-
-    const commandClause = formatReminderLabelValue(tt('命令'), commandPreview);
-    const reservedSuffixBudget = getReminderSentenceSeparator().length + commandClause.length;
-    const remainingSummaryBudget = STATUS_REMINDER_MAX_BODY_CHARS - reservedSuffixBudget;
-    if (remainingSummaryBudget <= 0) {
-        return truncateNotificationText(commandClause, STATUS_REMINDER_MAX_BODY_CHARS);
-    }
-
-    const summaryPreview = truncateNotificationText(summary, remainingSummaryBudget);
-    return appendReminderClause(summaryPreview, commandClause);
-}
-
-function buildWaitingUserQuestionBody(
-    question: NonNullable<SessionTaskSnapshot['pendingUserQuestion']>
-): string {
-    const summary = summarizeNotificationText(question.questions[0]?.question);
-    if (summary) {
-        return summary;
-    }
-
-    return tt('任务正在等待用户回复');
-}
-
-/**
- * 根据任务快照构建状态提醒负载。
- * 仅在 completed、failed、waiting_approval 三种状态下生成提醒，其余返回 null。
- */
 export function buildSessionStatusReminder(
     snapshot: SessionTaskSnapshot
 ): SessionStatusReminderPayload | null {
